requirements.txt
aioredis==1.3.1
alembic==1.10.3
asyncpg==0.27.0
fastapi==0.95.1
python-dotenv==1.0.0
fastapi-cache2==0.2.1
fastapi-users==10.4.2
fastapi-users-db-sqlalchemy==5.0.0
redis==4.5.5
gunicorn
celery
flower
httpx